removed changed from every control, moved it to base class.
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@2481 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ menu/map_picker.h
@@ -24,7 +24,7 @@ private:
 	void loadScreenshot();
 	void scan(const std::string &dir);
 
-	size_t _index;
+	int _index;
 	typedef std::vector<MapDesc > MapList;
 	MapList _maps;
